chore: Keep PR builds for at least 5 days.
It's too easy to lose track of older builds when we are removing all
PR `PipelineActivity`s after 48 hours. So let's bump that to 120 - if
there are performance problems due to the increase in CRDs retained (I
don't think there will be, but...), I'll improve `jx gc activity` to
be able to say "delete the `Pipeline` and friends after X duration,
but keep the `PipelineActivity` for Y duration".

Also add Hardy to OWNERS.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>