chore: Switch to Lighthouse post-log link and retest fixes
With https://github.com/jenkins-x/jx/issues/6791, https://github.com/jenkins-x/jx/issues/6792, https://github.com/jenkins-x/jx/issues/6794, and https://github.com/jenkins-x/jx/issues/6796 fixed since yesterday's attempt at switching weasel to Lighthouse, it's time for another attempt.

As with the previous attempt, the assumption is that issues will be found and we'll want to roll back to Prow. Which is fine!

This does differ from the previous attempt in that I've added the new `--job-url-base=...` argument to the build controller.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>