chore: Add new lighthouse-config library repository
https://github.com/jenkins-x/lighthouse-config - I'm going to be
switching jx and lighthouse to depend on this rather than jx
depending on Prow and Lighthouse having its own copy.

This will allow us to start adding things to the config without
needing a circular dependency (Lighthouse already depends on jx, so we
don't want jx to depend on Lighthouse!).

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>